factor out helper functions into seperate module
